+++ lib/trainsh.rb
@@ -3,3 +3,6 @@ require_relative 'trainsh/constants'
 require_relative 'trainsh/errors'
 # require_relative 'trainsh/log'
 require_relative 'trainsh/version'
+
+require_relative 'trainsh/detectors/target/env'
+require_relative 'trainsh/detectors/target/kitchen'