+++ lib/trainsh/cli.rb
@@ -4,14 +4,12 @@ require_relative 'session'
 require_relative 'mixin/builtin_commands'
 require_relative 'mixin/file_helpers'
 require_relative 'mixin/sessions'
-
-# TODO
-# require_relative 'detectors/target/env.rb'
-# require_relative 'detectors/target/kitchen.rb'
+require_relative 'mixin/shell_output'
 
 require 'colored'
 require 'fileutils'
 require 'readline'
+require 'rubygems'
 require 'train'
 require 'thor'
 
@@ -19,7 +17,6 @@ module TrainSH
   class Cli < Thor
     include Thor::Actions
     check_unknown_options!
-    # add_runtime_options!
 
     def self.exit_on_failure?
       true
@@ -34,10 +31,14 @@ module TrainSH
     EXIT_COMMANDS = %w[!!! exit quit logout disconnect].freeze
     INTERACTIVE_COMMANDS = %w[more less vi vim nano].freeze
 
+    NON_OS_TRANSPORTS = %w[aws core kubernetes azure pgsql vsphere vault digitalocean rest].freeze
+    CORE_TRANSPORTS = %w[docker ssh].freeze
+
     no_commands do
       include TrainSH::Mixin::BuiltInCommands
       include TrainSH::Mixin::FileHelpers
       include TrainSH::Mixin::Sessions
+      include TrainSH::Mixin::ShellOutput
 
       def __disconnect
         session.close
@@ -70,12 +71,19 @@ module TrainSH
       end
 
       def target_detectors
-        Dir[File.join(__dir__, 'lib', '*.rb')].sort.each { |file| require file }
-
         TrainSH::Detectors::TargetDetector.descendants
       end
 
+      def detect_target
+        target_detectors.detect(&:url).url
+      end
+
       def execute(input)
+        if input == '?'
+          execute_builtin 'help'
+          return
+        end
+
         case input[0]
         when '.'
           execute_locally input[1..]
@@ -142,7 +150,7 @@ module TrainSH
       end
 
       def prompt
-        exitcode = current_session.exitcode
+        exitcode = current_session.exitcode || 0
         exitcode_prefix = exitcode.zero? ? 'OK '.green : format('E%02d ', exitcode).red
 
         format(::TrainSH::PROMPT,
@@ -159,7 +167,7 @@ module TrainSH
 
         choices.concat(builtin_commands.map { |cmd| "!#{cmd.tr('_', '-')}" })
         choices.concat(sessions.map { |session_id| "@#{session_id}" })
-        choices.concat %w[!!!]
+        choices.concat %w[!!! ?]
 
         choices.filter { |choice| choice.start_with? partial }
       end
@@ -175,14 +183,55 @@ module TrainSH
       #
       #   Logger.const_get(l.upcase)
       # end
+
+      def local_gems
+        Gem::Specification.sort_by { |g| [g.name.downcase, g.version] }.group_by(&:name)
+      end
     end
 
     # class_option :log_level, desc: "Log level", aliases: "-l", default: :info
     class_option :messy, desc: 'Skip deletion of temporary files for speedup', default: false, type: :boolean
 
     desc 'connect URL', 'Connect to a destination interactively'
+    long_desc <<-DESC
+      Create an interactive shell session with the remote system. The specified URL has to match the
+      chosen transport plugin.
 
-    def connect(url)
+      If no URL was given, possible targets are detected from the environment variable TARGET or any
+      existing Test Kitchen instances (max: 1).
+
+      URL Examples:
+        docker://d9443b195d16
+        local://
+        ssh://user@remote.example.com
+        winrm://Administrator:PASSWORD@10.2.42.1
+
+      URL Examples from non-standard transports:
+        aws-ssm://i-1234567890ab
+        serial://dev/ttyUSB1/9600
+        telnet://127.0.0.1
+        vsphere-gom://Administrator@vcenter.server/virtual.machine
+
+      Every transport has its own, proprietary options which can currently only be added as URL
+      query parameters:
+        ssh://user@remote.example.com?key_files=/home/ubuntu/test.pem
+
+      Passwords currently have to be part of the URL.
+    DESC
+    def connect(url = nil)
+      # TODO: Pass options to `use_session`
+      unless url
+        show_message 'No URL given, trying to detect ...'
+        url = detect_target
+
+        show_message "Detected URL to be #{url}" if url
+      end
+
+      unless url
+        show_error 'No target could be detected'
+        exit
+      end
+
       exit unless use_session(url)
 
       say format('Connected to %<url>s', url: session.url).bold
@@ -227,14 +276,19 @@ module TrainSH
 
         execute input
       end
+    rescue Interrupt
+      show_error 'Interrupted execution'
     end
 
     # desc 'copy FILE/DIR|URL FILE/DIR|URL', 'Copy files or directories'
     # def copy(url_or_file, url_or_file)
-    #   # TODO?
+    #   # TODO
     # end
 
     desc 'detect URL', 'Retrieve remote OS and platform information'
+    long_desc <<~DESC
+      Detect remote OS via Train. Uses the same schema as URLs for `connect`.
+    DESC
     def detect(url)
       exit unless use_session(url)
       __detect
@@ -249,13 +303,10 @@ module TrainSH
 
     desc 'list-transports', 'List available transports'
     def list_transports
-      # TODO: Train only lazy-loads and does not have a full registry
-      #       https://github.com/inspec/train/blob/be90ca53ea1c1e8aa7439c504fbee86f4b399d83/lib/train.rb#L38-L61
+      installed = local_gems.select { |name| name.start_with? 'train-' }.keys.map { |name| name.delete_prefix('train-') }
+      transports = installed - NON_OS_TRANSPORTS + CORE_TRANSPORTS
 
-      # TODO: Filter for only "OS" transports as well
-      transports = %w[local ssh winrm docker]
-
-      say "Available transports: #{transports.sort.join(', ')}"
+      say "Installed transports: #{transports.sort.join(', ')}"
     end
   end
 end