+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## Version 0.3.0
+
+- Add auto-detection of targets from environment (variable, test-kitchen)
+- Add `!copy` for cross-session copying
+- Add `!help` and `?` for usage in shell
+- Add detailed help for commands
+- Add output on copying files cross-session and editing/saving
+- Add devcontainer for VSCode
+- Change errors to be red, Shell built-in to yellow
+- Fix handling of URL encoded parameters
+- Fix `list-transports` to output all installed non-API transports
+- Fix crashes on unexpected input in shell
+- Fix exception handling in various places
+- Fix missing binary in Gem
+
 ## Version 0.2.0
 
 - Renamed to TrainSH